+++ include/world.hpp
@@ -5,7 +5,6 @@
 #include "player_character.hpp"
 #include "stats.hpp"
 #include <memory>
-#include <stdint.h>
 #include <string>
 #include <vector>
 
@@ -29,6 +28,5 @@ class World {
     std::vector<std::vector<Entity*>> overworld_occupants_;
 
     void populate_overworld();
-    void move_entity(Entity* entity, int_fast8_t new_x_pos,
-                     int_fast8_t new_y_pos) noexcept;
+    void move_entity(Entity* entity, int new_x_pos, int new_y_pos) noexcept;
 };
